+++ packages/api-evm/source/actions/index.ts
@@ -1,5 +1,5 @@
-export * from "./call.js";
 export * from "./eth-block-number.js";
+export * from "./eth-call.js";
 export * from "./eth-get-balance.js";
 export * from "./net-listening.js";
 export * from "./net-peer-count.js";
